Refactor for modularity, DRY out some code
