+++ react/new/src/components/Button/Button.tsx
@@ -1,18 +1,11 @@
-import { useTheme } from '../../contexts/ThemeContext'
 import styles from './button.module.scss'
-import { ButtonHTMLAttributes, CSSProperties, ReactNode } from 'react'
-
-export type ButtonVariant = 'primary' | 'secondary' | 'text'
+import { ButtonHTMLAttributes, ReactNode } from 'react'
+import { ButtonVariant, useButtonStyles } from './useButtonStyles'
 
 export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   variant?: ButtonVariant
   className?: string
-  style?: CSSProperties
   children?: ReactNode
-}
-
-type CSSPropertiesWithVars = CSSProperties & {
-  [key: `--${string}`]: string
 }
 
 export function Button({
@@ -22,76 +15,15 @@ export function Button({
   children,
   ...props
 }: ButtonProps) {
-  const {
-    theme,
-    getHeaderBackgroundColor,
-    getAccessibleLinkColor,
-    getAccessibleActiveLinkColor,
-  } = useTheme()
-
-  const getButtonStyles = () => {
-    const baseStyle: CSSPropertiesWithVars = {
-      ...(style as CSSPropertiesWithVars),
-    }
-
-    if (variant === 'primary') {
-      const bgColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const textColor = getAccessibleLinkColor()
-      const hoverColor = getAccessibleActiveLinkColor()
-
-      baseStyle.backgroundColor = bgColor
-      baseStyle.color = textColor
-      baseStyle.border = 'none'
-
-      baseStyle['--hover-bg-color'] = hoverColor
-      baseStyle['--hover-text-color'] = '#fff'
-      baseStyle['--primary-color'] = hoverColor
-    } else if (variant === 'secondary') {
-      const borderColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const textColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const hoverColor = getAccessibleActiveLinkColor()
-
-      baseStyle.backgroundColor = '#fff'
-      baseStyle.color = textColor
-      baseStyle.border = `1px solid ${borderColor}`
-
-      baseStyle['--hover-bg-color'] = hoverColor
-      baseStyle['--hover-text-color'] = '#fff'
-      baseStyle['--hover-border-color'] = hoverColor
-      baseStyle['--primary-color'] = hoverColor
-    } else if (variant === 'text') {
-      const textColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const hoverColor = getAccessibleActiveLinkColor()
-
-      baseStyle.color = textColor
-
-      baseStyle['--hover-text-color'] = hoverColor
-      baseStyle['--primary-color'] = hoverColor
-    }
-
-    return baseStyle
-  }
+  const { styles: buttonStyles, className: variantClass } = useButtonStyles(
+    variant,
+    style
+  )
 
   return (
     <button
-      className={`${styles[variant]} ${className || ''}`}
-      style={getButtonStyles()}
+      className={`${styles[variantClass]} ${className || ''}`}
+      style={buttonStyles}
       type='button'
       {...props}
     >
