+++ react/new/src/components/Button/LinkButton.tsx
@@ -1,20 +1,13 @@
-import { useTheme } from '../../contexts/ThemeContext'
 import styles from './button.module.scss'
-import { AnchorHTMLAttributes, CSSProperties, ReactNode } from 'react'
-
-export type LinkButtonVariant = 'primary' | 'secondary' | 'text'
+import { AnchorHTMLAttributes, ReactNode } from 'react'
+import { ButtonVariant, useButtonStyles } from './useButtonStyles'
 
 export interface LinkButtonProps
   extends AnchorHTMLAttributes<HTMLAnchorElement> {
-  variant?: LinkButtonVariant
+  variant?: ButtonVariant
   className?: string
-  style?: CSSProperties
   children?: ReactNode
   href: string
-}
-
-type CSSPropertiesWithVars = CSSProperties & {
-  [key: `--${string}`]: string
 }
 
 export function LinkButton({
@@ -25,76 +18,15 @@ export function LinkButton({
   href,
   ...props
 }: LinkButtonProps) {
-  const {
-    theme,
-    getHeaderBackgroundColor,
-    getAccessibleLinkColor,
-    getAccessibleActiveLinkColor,
-  } = useTheme()
-
-  const getLinkStyles = () => {
-    const baseStyle: CSSPropertiesWithVars = {
-      ...(style as CSSPropertiesWithVars),
-    }
-
-    if (variant === 'primary') {
-      const bgColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const textColor = getAccessibleLinkColor()
-      const hoverColor = getAccessibleActiveLinkColor()
-
-      baseStyle.backgroundColor = bgColor
-      baseStyle.color = textColor
-      baseStyle.border = 'none'
-
-      baseStyle['--hover-bg-color'] = hoverColor
-      baseStyle['--hover-text-color'] = '#fff'
-      baseStyle['--primary-color'] = hoverColor
-    } else if (variant === 'secondary') {
-      const borderColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const textColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const hoverColor = getAccessibleActiveLinkColor()
-
-      baseStyle.backgroundColor = '#fff'
-      baseStyle.color = textColor
-      baseStyle.border = `1px solid ${borderColor}`
-
-      baseStyle['--hover-bg-color'] = hoverColor
-      baseStyle['--hover-text-color'] = '#fff'
-      baseStyle['--hover-border-color'] = hoverColor
-      baseStyle['--primary-color'] = hoverColor
-    } else if (variant === 'text') {
-      const textColor =
-        theme.type === 'team' && theme.team
-          ? getHeaderBackgroundColor()
-          : '#000'
-
-      const hoverColor = getAccessibleActiveLinkColor()
-
-      baseStyle.color = textColor
-
-      baseStyle['--hover-text-color'] = hoverColor
-      baseStyle['--primary-color'] = hoverColor
-    }
-
-    return baseStyle
-  }
+  const { styles: buttonStyles, className: variantClass } = useButtonStyles(
+    variant,
+    style
+  )
 
   return (
     <a
-      className={`${styles[variant]} ${className || ''}`}
-      style={getLinkStyles()}
+      className={`${styles[variantClass]} ${className || ''}`}
+      style={buttonStyles}
       href={href}
       {...props}
     >
