+++ react/new/src/components/Button/index.ts
@@ -1,2 +1,3 @@
 export * from './Button'
 export * from './LinkButton'
+export * from './useButtonStyles'
